Added Mean Time to Failure (MTTF)
Added the MTTF calculation for both an individual component and the whole system. To determine MTTF, the failure state has to become an absorbing state; the new .csv with abs in the title contains the values of the new Q matrix. Also, for the MTTF calculation, you need to have a much longer period. I found that 300 years is pretty accurate, while 100 years gives a good idea of the MTTF that we want.